session-cards #1 implementation
Fixed minor bugs and visual updates for navbar, hero, sessions

Todo:

1. Fix mobile navbar links
2. Fix news section on landing page

+++ backend/internal/models/constructor_standing.go
@@ -8,12 +8,12 @@ import (
 type ConstructorStanding struct {
 	ID          uuid.UUID `gorm:"type:uuid;primary_key;default:gen_random_uuid()" json:"id"`
 	TeamID      uuid.UUID `gorm:"type:uuid;not null" json:"team_id"`
-	Season      int8      `gorm:"not null" json:"season"`
-	RoundNumber int8      `gorm:"not null" json:"round_number"`
+	Season      int16     `gorm:"not null" json:"season"`
+	RoundNumber int16     `gorm:"not null" json:"round_number"`
 
-	Position int8    `gorm:"not null" json:"position"`
+	Position int16   `gorm:"not null" json:"position"`
 	Points   float64 `gorm:"type:numeric(10,2);not null" json:"points"`
-	Wins     int8    `gorm:"default:0" json:"wins"`
+	Wins     int16   `gorm:"default:0" json:"wins"`
 
 	// Relationships
 	Team *Team `gorm:"foreignKey:TeamID;references:ID" json:"team,omitempty"`
